[DEV] Implement the `.chef` configuration file.
At startup `chef` looks for a `.chef` file in the current directory and
the recursive parents up to the root of the file system.

`.chef` is a JSON formatted file with currently only one useful object: `menu`.

The `chef init <menu-name>` command prepares the `.chef` file.

For example:

```
$ chef init pi3-sample-menu.json
```

produces the `.chef` file:

```
{"menu": "pi3-sample-menu.json"}
```

Then you don't need to repeat the menu name for the upcoming commands
(`prepare`, `build`).

On the other hand, indicating a menu file name as argument to the other commands
takes precedence over the `.chef` content.