coocker/cooker.py: introduce configuration template (not buildable)
This change allow to define configuration template (not buildable) to
be used with 'inherit' keyword.

Distinguish between buildable and not-buildable configurations using
the same method as for gitlab-ci. If the configuration name start with
'.' it's a template and must not be build.

If the configuration doesn't start with '.' make sure that 'target' is
defined by one of the ancestors.

Implement #81

Signed-off-by: Romain Naour <romain.naour@smile.fr>