cooker: add --keepgoing option for build command
bitbake provide the option '-k' or '--continue' to continue as much as
possible after an error.

This is useful for the first build for a new target or after a layer
update when we can expect having some build issues.

Add a new test that check if "-k" is on the bitbake command line if
--keepgoing is used.

Signed-off-by: Romain Naour <romain.naour@gmail.com>