[DEV] menu path can be absolute or relative
When the menu path given in the command init (or cook) started with '/', its
absolute path is stored (after symlinks resolution). If not its path relative to
the .cookerconfig is stored (after symlinks resolution).

The aim is to enable to initialize and download the sources in a host machine
and build in a container which may have a different location.

Add a test in init for relative paths.