[FIX] extrauser command with hashed passwords
Fix an error in the usermod command. On recent shadow version, `useradd` and
`usermod` option `-P` is for ``--prefix` and the build fails.

Passwords must be given with the option `-p` with hashed passwords. Because it
is a bit complicated to escape correctly \", we remove this feature from almost
all menus and create `riscv-menu-extrauser.json` as an example.