test: basic: dry-run: Modify output.ref
The new method for obtaining the path to the template file utilizes the poky
version. However, in order to utilize this method, we must first clone the poky
repository. During a dry-run, the repository is not cloned, resulting in a
return value of "None" as the correct path cannot be found.

To ensure the accuracy of the path, two additional tests will be implemented.

Signed-off-by: Ever ATILANO <ever.atilano@smile.fr>
Reviewed-by: Yoann Congal <yoann.congal@smile.fr>