[DEV] Add support for executing a command in a poky-initialized shell
Updated the shell command to support executing a command within the poky-initialized shell.
Introduced the cmd argument, which allows users to specify a command to be executed.
When the cmd argument is provided, the specified command is executed within the shell environment.

Example: cooker shell <cooker-build>  'bitbake -c cve_check <package-name>'

Signed-off-by: Lamine REHAHLIA <lamine.rehahlia@smile.fr>
Reviewed-by: Yoann CONGAL <yoann.congal@smile.fr>