Fix SHELL=/bin/sh=dash support (#159)
* [FIX] shell: Use "." instead of "source"

The "." command is more portable as it is defined in POSIX[0], whereas
the "source" command is a bash (and some other shells) extensions.

In particular, "source" is not supported by dash, the default /bin/sh on
Debian 12.

[0]: https://pubs.opengroup.org/onlinepubs/9799919799/utilities/V3_chap02.html#dot

Signed-off-by: Yoann Congal <yoann.congal@smile.fr>

* [FIX] shell: Dash support for init_script

Dash does not support passing arguments to the "." command. The default
poky ./oe-init-build-env has this comment:
  # Normally this is called as '. ./oe-init-build-env <builddir>'
  #
  # This works in most shells (not dash), but not all of them pass the arguments
  # when being sourced.  To workaround the shell limitation use "set <builddir>"
  # prior to sourcing this script.

Implement the "set" workaround for dash support.

Also, change current working directory to match oe-init-build-env
expectation of being sourced from its base directory in dash.

Signed-off-by: Yoann Congal <yoann.congal@smile.fr>

---------

Signed-off-by: Yoann Congal <yoann.congal@smile.fr>