[FIX] Unsort BBLAYERS for pylib import
Since BitBake added addpylib support[0], layers that include a lib
directory need to use the addpylib directive in their layer.conf file.
For example:

  addpylib ${LAYERDIR}/lib oeqa

Lib directories are added to sys.path in BBLAYERS order, which is
currently sorted during the bblayers.conf generation. This can raise a
ModuleNotFoundError if a custom module tries to import a sub-module
that appears later in the sys.path.

Removing the BBLAYERS sort in bblayers.conf allows users to manually
order layers via the menu.json file, resolving this dependency issue.

[0]: https://github.com/yoctoproject/poky/commit/42e8b2682f56f450c1258956213198052d6bcb96

Signed-off-by: Pierre-Loup GOSSE <pierre-loup.gosse@smile.fr>