[DEV] Add the `NoPoky` base-distribution. (#162)
* [DEV] Add the `NoPoky` base-distribution.

As the Yocto Project maintainers have announced that they
are discontinuing maintenance of the Poky repository, we
need to adapt Cooker, which relied heavily on Poky.

This base distribution can be used in a menu by adding
the line
` "base-distribution": "NoPoky" `

I assume that in a few months we will be able to switch
to the "NoPoky" distribution as the default distro and use
`"base-distribution": "Poky"` to build older versions.

* Modify distro.py (required by ruff

* Modify distro.py (required by mypy)

* [FIX] Avoid a warning about duplicate inclusion of default distro.

* [DEV] Add a simple Qemu menu that does not use Poky repo.